requirements.txt
fastapi
torch
transformers
flask
flask_cors
uvicorn
pillow
python-multipart
numpy
opencv-python
pyvips
albumentations
scikit-learn
tensorboard
accelerate>=0.26.0